pyproject.toml
[tool.poetry]
name = "cardapiorapido-project"
version = "0.1.0"
description = ""
authors = ["Rip4568 <63131796+Rip4568@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
python-decouple = "^3.8"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
django-extensions = "^3.2.3"
slugify = "^0.0.1"
django-filter = "^24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
